requirements.txt
transformers>=5.0.0
pillow
